# apis/openapi.yaml
openapi: 3.1.0
info:
  title: Network Inventory Planning API
  version: 1.0.0
  contact:
    name: warehouse-systems
    url: https://github.com/IQVO/network-inventory-planning
  description: |
    Advisory network inventory balancing for the warehouse-systems fleet.

    Network Inventory Planning projects sibling-context facts (site capability,
    site-SKU demand, published capacity) into local read models and proposes
    inter-warehouse transfers. It never mutates physical stock: approval,
    reservation and physical execution live in their owning contexts.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/transfers/{id}:cancel:
    post:
      operationId: cancelTransfer
      tags: [transfers]
      summary: Cancel a transfer that has not yet been released (operator action).
      description: |
        The operator's PRE-RELEASE cancel (ADR 0011, ADR 0003). Legal only
        while the transfer is DRAFT, PROPOSED, APPROVED or ALLOCATING; once
        an origin reservation exists (ALLOCATED) or the transfer is past it,
        the request is a 409 `transfer-not-cancellable` and nothing changes.
        The new CANCELLED state, its audit entry (cause = `reason`) and the
        TransferStateAdvanced analytics occurrence commit in one
        transaction. No integration event is emitted. Idempotent: cancelling
        an already CANCELLED transfer is a 200 no-op that returns the
        transfer unchanged. A blank `reason` is a 422 `invalid-cancellation`;
        an unknown id is a 404 `transfer-not-found`; 503 problem+json when the
        instance runs without persistence and the outbox. REST only: the MCP
        surface stays read-only.
      parameters:
        - name: id
          in: path
          required: true
          description: The transfer id (the transferId returned by approval).
          schema:
            type: string
            minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelTransferRequest'
      responses:
        '200':
          description: The transfer, now CANCELLED (also returned for an already CANCELLED transfer).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferView'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
...
components:
...
  responses:
    Problem:
      description: RFC 7807 problem+json error.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          description: Stable machine-readable problem identifier.
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
    SiteID:
      type: string
      minLength: 1
      description: Facility-layout Site code identifying one warehouse.
    SKU:
      type: string
      minLength: 1
...
    TransferState:
      type: string
      enum: [DRAFT, PROPOSED, APPROVED, ALLOCATING, ALLOCATED, PICKED, IN_TRANSIT, ARRIVED, RECEIVED, UNFULFILLABLE, CANCELLED]
      description: |
        Saga lifecycle state. RECEIVED, UNFULFILLABLE and CANCELLED are
        terminal; every other state is waiting for a next step.
    TransferView:
      type: object
      required: [id, state, sku, quantity, originSiteId, destinationSiteId, policyVersion, expiresAt, createdAt, updatedAt, version]
      properties:
        id:
          type: string
          description: The transfer (saga aggregate) id.
        state:
          $ref: '#/components/schemas/TransferState'
        sku:
          $ref: '#/components/schemas/SKU'
        quantity:
          type: integer
          description: Units planned to move.
        pickedQuantity:
          type: integer
          description: Units the origin pick actually picked; present from PICKED onwards (may be short of quantity).
        originSiteId:
          $ref: '#/components/schemas/SiteID'
        destinationSiteId:
          $ref: '#/components/schemas/SiteID'
        policyVersion:
          type: string
        reservationId:
          type: string
          description: inventory-storage's reservation id, present once ALLOCATED.
        rejectionReason:
          type: string
          enum: [ORIGIN_SITE_UNKNOWN, INSUFFICIENT_USABLE, IDEMPOTENCY_CONFLICT]
          description: The closed rejection reason, present once UNFULFILLABLE.
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          description: Time of the last state transition; a non-terminal transfer whose updatedAt is old is stuck.
        version:
          type: integer
          format: int64
          description: Optimistic-concurrency version (the number of persisted state transitions).
...
    CancelTransferRequest:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 1
          pattern: '\S'
          description: The operator's justification; required and not blank. It becomes the cause of the TransferCancelled audit entry.
...